+++ kanidm_tools/src/cli/raw.rs
@@ -1,4 +1,23 @@
+use crate::common::CommonOpt;
+use kanidm_proto::v1::{Entry, Filter, Modify, ModifyList};
 use std::collections::BTreeMap;
+use structopt::StructOpt;
+
+use std::error::Error;
+use std::fs::File;
+use std::io::BufReader;
+use std::path::Path;
+use std::path::PathBuf;
+
+use serde::de::DeserializeOwned;
+
+fn read_file<T: DeserializeOwned, P: AsRef<Path>>(path: P) -> Result<T, Box<dyn Error>> {
+    let f = File::open(path)?;
+    let r = BufReader::new(f);
+
+    let t: T = serde_json::from_reader(r)?;
+    Ok(t)
+}
 
 #[derive(Debug, StructOpt)]
 pub struct FilterOpt {
@@ -39,6 +58,15 @@ pub enum RawOpt {
 }
 
 impl RawOpt {
+    pub fn debug(&self) -> bool {
+        match self {
+            RawOpt::Search(sopt) => sopt.commonopts.debug,
+            RawOpt::Create(copt) => copt.commonopts.debug,
+            RawOpt::Modify(mopt) => mopt.commonopts.debug,
+            RawOpt::Delete(dopt) => dopt.commonopts.debug,
+        }
+    }
+
     pub fn exec(&self) -> () {
         match self {
             RawOpt::Search(sopt) => {
@@ -54,7 +82,7 @@ impl RawOpt {
             RawOpt::Create(copt) => {
                 let client = copt.commonopts.to_client();
                 // Read the file?
-                match copt.file {
+                match &copt.file {
                     Some(p) => {
                         let r_entries: Vec<BTreeMap<String, Vec<String>>> = read_file(p).unwrap();
                         let entries = r_entries.into_iter().map(|b| Entry { attrs: b }).collect();
@@ -68,7 +96,7 @@ impl RawOpt {
             RawOpt::Modify(mopt) => {
                 let client = mopt.commonopts.to_client();
                 // Read the file?
-                match mopt.file {
+                match &mopt.file {
                     Some(p) => {
                         let filter: Filter = serde_json::from_str(mopt.filter.as_str()).unwrap();
                         let r_list: Vec<Modify> = read_file(p).unwrap();
@@ -88,4 +116,3 @@ impl RawOpt {
         }
     }
 }
-
